+++ js/popup.js
@@ -1,5 +1,9 @@
 var bg = chrome.extension.getBackgroundPage();
 
+// window.CollecUsage(function(){
+//     localStorage.clear();
+// });
+
 chrome.tabs.getSelected(null,
 function(tab) {
     chrome.extension.sendMessage({
@@ -10,48 +14,41 @@ function(tab) {
         var display = document.getElementById('app_list');
 
         var apps = response && response.apps ? response.apps: {};
-        var html = '';
 
         var appinfo = bg.appinfo;
-        var count = 0;
 
         for (var appid in apps) {
             app = appinfo[appid] ? appinfo[appid] : {};
 
-            // i'm lazy to fill all kind of the information :(
             if (!app.title) app.title = appid;
             if (!app.url) app.url = appinfo[''].url.replace('%s', appid); // it's google one
-            if (!app.icon) app.icon = appinfo[''].icon;
+            // if (!app.icon) app.icon = appinfo[''].icon;
 
             if (apps[appid] != "-1") {
-                app.title = appid + ' ' + apps[appid]
+                app.title = appid + ' <span class="lib_version">' + apps[appid] + '</span>';
             }
 
-            // use DOM to avoid error
             var link = document.createElement('a');
-            var icon = document.createElement('img');
-
             link.target = "_blank";
-            link.title = app.title;
+            // link.title = app.title;
             link.href = app.url;
 
-            icon.alt = app.title;
-            icon.width = 16;
-            icon.height = 16;
-            icon.src = "apps/" + app.icon;
+            if (app.icon !== undefined) {
+                var icon = document.createElement('img');
+                icon.width = 16;
+                icon.height = 16;
+                icon.src = "apps/" + app.icon;
+            } else {
+                var icon = document.createElement('span');
+                icon.className = "icon";
+            }
+
+            var text = document.createElement('span');
+            text.innerHTML = app.title;
 
             link.appendChild(icon);
+            link.appendChild(text);
             display.appendChild(link);
-
-            count++;
         }
-
-        if (count < 8) // correct the width for better view
-        {
-            display.style.width = (count * 20) + "px";
-        } else {
-            display.style.width = "160px";
-        }
-
     });
 });